+++ src/Integrations/CRM/Freshdesk.php
@@ -83,7 +83,7 @@ class Freshdesk extends AbstractCRMIntegration
      *
      * @return bool
      */
-    public function pushObject(array $keyValueList): bool
+    public function pushObject(array $keyValueList, $formFields = null): bool
     {
         $requestType = 'json';